requirements.txt
# Production Dependencies
# Core
python-telegram-bot[webhooks,rate-limiter]==20.7
pyrogram==2.0.106
telethon==1.34.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Cache & Queue
redis==5.0.1
celery==5.3.4
flower==2.0.1
kombu==5.3.4

# Storage
minio==7.2.0
boto3==1.34.11

# Security
cryptography==41.0.7
pyjwt==2.8.0
passlib==1.7.4
python-jose==3.3.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
elasticsearch==8.11.1

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Media Processing
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
Pillow==10.2.0
playwright==1.40.0
beautifulsoup4==4.12.2
mutagen==1.47.0

# Utils
httpx==0.26.0
aiohttp==3.9.1
python-dotenv==1.0.0
babel==2.14.0
pytz==2023.3

# Payment
stripe==7.10.0
coinbase-commerce==1.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.0.0

# Development
black==23.12.1
ruff==0.1.9
mypy==1.8.0
pre-commit==3.6.0